@@ -1,10 +1,9 @@
-@page "/myorders"
+﻿@page "/myorders"
 @inject HttpClient HttpClient
 @inject NavigationManager NavigationManager
 
-
 <div class="main">
-    @if (ordersWithStatus == null)
+    @if (ordersWithStatus is null)
     {
         <text>Loading...</text>
     }
@@ -27,13 +26,14 @@
                         <strong>£@item.Order.GetFormattedTotalPrice()</strong>
                     </div>
                     <div class="col">
-                        Status <strong>@item.StatusText</strong>
+                        Status: <strong>@item.StatusText</strong>
                     </div>
                     @if (@item.StatusText != "Delivered")
                     {
                         <div class="col flex-grow-0">
                             <a href="myorders/@item.Order.OrderId" class="btn btn-success">
-                                Track &gt;</a>
+                                Track &gt;
+                            </a>
                         </div>
                     }
                 </div>
@@ -43,10 +43,9 @@
 </div>
 
 @code {
-    List<OrderWithStatus> ordersWithStatus = new List<OrderWithStatus>();
+    List<OrderWithStatus>? ordersWithStatus = new();
 
-    protected override async Task OnParametersSetAsync()
-    {
-        ordersWithStatus = await HttpClient.GetFromJsonAsync<List<OrderWithStatus>>($"{NavigationManager.BaseUri}orders");
-    }
+    protected override async Task OnParametersSetAsync() =>
+        ordersWithStatus = await HttpClient.GetFromJsonAsync<List<OrderWithStatus>>(
+            $"{NavigationManager.BaseUri}orders");
 }
