@@ -2,23 +2,19 @@
 @inject NavigationManager NavigationManager
 @inject HttpClient HttpClient
 
-
 <div class="main">
     @if (invalidOrder)
     {
         <h2>Order not found</h2>
-        <p>We;re sorry but this order no longer exists.</p>
+        <p>We're sorry but this order no longer exists.</p>
     }
-    else if (orderWithStatus == null)
+    else if (orderWithStatus is null)
     {
         <div class="track-order">
             <div class="track-order-title">
                 <h2>
                     <text>Loading...</text>
                 </h2>
-                <p class="ml-auto mb-0">
-                    ...
-                </p>
             </div>
         </div>
     }
@@ -40,7 +36,7 @@
                         <p>
                             <strong>
                                 @(pizza.Size)"
-                                @pizza.Special.Name
+                                @pizza.Special!.Name
                                 (£@pizza.GetFormattedTotalPrice())
                             </strong>
                         </p>
@@ -53,13 +49,16 @@
 
 @code {
     [Parameter] public int OrderId { get; set; }
-    OrderWithStatus orderWithStatus;
+
+    OrderWithStatus? orderWithStatus;
     bool invalidOrder = false;
+
     protected override async Task OnParametersSetAsync()
     {
         try
         {
-            orderWithStatus = await HttpClient.GetFromJsonAsync<OrderWithStatus>($"{NavigationManager.BaseUri}orders/{OrderId}");
+            orderWithStatus = await HttpClient.GetFromJsonAsync<OrderWithStatus>(
+                $"{NavigationManager.BaseUri}orders/{OrderId}");
         }
         catch (Exception ex)
         {
