@@ -1,23 +1,22 @@
-using Microsoft.EntityFrameworkCore;
-
 namespace BlazingPizza.Data;
 
-public class PizzaStoreContext : DbContext
+public sealed class PizzaStoreContext(DbContextOptions options)
+    : DbContext(options)
 {
-    public PizzaStoreContext(DbContextOptions options) : base(options)
-    {
+    public DbSet<Order> Orders => Set<Order>();
 
-    }
+    public DbSet<Pizza> Pizzas => Set<Pizza>();
 
-    public DbSet<Order> Orders { get; set; }
-    public DbSet<Pizza> Pizzas { get; set; }
-    public DbSet<PizzaSpecial> Specials { get; set; }
-    public DbSet<PizzaSpecial> Toppings { get; set; }
+    public DbSet<PizzaSpecial> Specials => Set<PizzaSpecial>();
+
+    public DbSet<Topping> Toppings => Set<Topping>();
+
     protected override void OnModelCreating(ModelBuilder modelBuilder)
     {
         base.OnModelCreating(modelBuilder);
 
-        // Configuring a many to many special -> topping relationship that is friendly for serialization
+        // Configuring a many-to-many special -> topping relationship
+        // that is friendly for serialization
         modelBuilder.Entity<PizzaTopping>().HasKey(pst => new { pst.PizzaId, pst.ToppingId });
         modelBuilder.Entity<PizzaTopping>().HasOne<Pizza>().WithMany(ps => ps.Toppings);
         modelBuilder.Entity<PizzaTopping>().HasOne(pst => pst.Topping).WithMany();
