@@ -1,18 +1,19 @@
 namespace BlazingPizza.Data;
 
-public static class SeedData
+internal static class SeedData
 {
-    public static void Initialize(PizzaStoreContext db)
+    internal static Task InitializeAsync(PizzaStoreContext db)
     {
-        var specials = new PizzaSpecial[] {
-            new PizzaSpecial()
+        PizzaSpecial[] specials =
+        [
+            new()
             {
                 Name = "Basic Cheese Pizza",
                 Description = "It's cheesy and delicious. Why wouldn't you want one?",
                 BasePrice = 9.99m,
                 ImageUrl = "img/pizzas/cheese.jpg",
             },
-            new PizzaSpecial()
+            new()
             {
                 Id = 2,
                 Name = "The Baconatorizor",
@@ -20,7 +21,7 @@ public static class SeedData
                 BasePrice = 11.99m,
                 ImageUrl = "img/pizzas/bacon.jpg",
             },
-            new PizzaSpecial()
+            new()
             {
                 Id = 3,
                 Name = "Classic pepperoni",
@@ -28,7 +29,7 @@ public static class SeedData
                 BasePrice = 10.50m,
                 ImageUrl = "img/pizzas/pepperoni.jpg",
             },
-            new PizzaSpecial()
+            new()
             {
                 Id = 4,
                 Name = "Buffalo chicken",
@@ -36,7 +37,7 @@ public static class SeedData
                 BasePrice = 12.75m,
                 ImageUrl = "img/pizzas/meaty.jpg",
             },
-            new PizzaSpecial()
+            new()
             {
                 Id = 5,
                 Name = "Mushroom Lovers",
@@ -44,7 +45,7 @@ public static class SeedData
                 BasePrice = 11.00m,
                 ImageUrl = "img/pizzas/mushroom.jpg",
             },
-            new PizzaSpecial()
+            new()
             {
                 Id = 7,
                 Name = "Veggie Delight",
@@ -52,7 +53,7 @@ public static class SeedData
                 BasePrice = 11.50m,
                 ImageUrl = "img/pizzas/salad.jpg",
             },
-            new PizzaSpecial()
+            new()
             {
                 Id = 8,
                 Name = "Margherita",
@@ -60,8 +61,8 @@ public static class SeedData
                 BasePrice = 9.99m,
                 ImageUrl = "img/pizzas/margherita.jpg",
             },
-        };
+        ];
         db.Specials.AddRange(specials);
-        db.SaveChanges();
+        return db.SaveChangesAsync();
     }
 }
