@@ -1,25 +1,22 @@
 ﻿@page "/"
+@using Microsoft.AspNetCore.Components.Web
 @namespace BlazingPizza.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 
 <!DOCTYPE html>
 <html lang="en">
-
 <head>
     <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
-    <base href="/" />
-    <link rel="icon" href="img/icon-512.png" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <base href="~/" />
     <link href="css/bootstrap/bootstrap.min.css" rel="stylesheet" />
-    <link href="css/site.css" rel="stylesheet" />
-    <title>Blazing Pizza</title>
+    <link href="css/site.css" rel="stylesheet" asp-append-version="true" />
+    <link href="BlazingPizza.styles.css" rel="stylesheet" />
+    <link rel="icon" type="image/png" href="favicon.png"/>
+    <component type="typeof(HeadOutlet)" render-mode="ServerPrerendered" />
 </head>
-
 <body>
-
-
     <component type="typeof(App)" render-mode="ServerPrerendered" />
-
 
     <div id="blazor-error-ui">
         <environment include="Staging,Production">
@@ -34,5 +31,4 @@
 
     <script src="_framework/blazor.server.js"></script>
 </body>
-
 </html>
