@@ -2,7 +2,7 @@
     <label>Name:</label>
     <div>
         <InputText @bind-Value="Address.Name" />
-        <ValidationMessage For="@(()=> Address.Name)" />
+        <ValidationMessage For="@(() => Address.Name)" />
     </div>
 </div>
 
@@ -10,38 +10,43 @@
     <label>Line 1:</label>
     <div>
         <InputText @bind-Value="Address.Line1" />
-        <ValidationMessage For="@(()=> Address.Line1)" />
+        <ValidationMessage For="@(() => Address.Line1)" />
     </div>
 </div>
+
 <div class="form-field">
     <label>Line 2:</label>
     <div>
         <InputText @bind-Value="Address.Line2" />
-        <ValidationMessage For="@(()=> Address.Line2)" />
+        <ValidationMessage For="@(() => Address.Line2)" />
     </div>
 </div>
+
 <div class="form-field">
     <label>City:</label>
     <div>
         <InputText @bind-Value="Address.City" />
-        <ValidationMessage For="@(()=> Address.City)" />
+        <ValidationMessage For="@(() => Address.City)" />
     </div>
 </div>
+
 <div class="form-field">
     <label>Region:</label>
     <div>
         <InputText @bind-Value="Address.Region" />
-        <ValidationMessage For="@(()=> Address.Region)" />
+        <ValidationMessage For="@(() => Address.Region)" />
     </div>
 </div>
+
 <div class="form-field">
     <label>Postal code:</label>
     <div>
         <InputText @bind-Value="Address.PostalCode" />
-        <ValidationMessage For="@(()=> Address.PostalCode)" />
+        <ValidationMessage For="@(() => Address.PostalCode)" />
     </div>
 </div>
 
 @code {
-    [Parameter] public Address Address { get; set; }
+    [Parameter, EditorRequired]
+    public required Address Address { get; set; }
 }
