@@ -3,21 +3,23 @@
 <div class="dialog-container">
     <div class="dialog">
         <div class="dialog-title">
-            <h2>@Pizza.Special.Name</h2>
+            <h2>@Pizza.Special!.Name</h2>
             @Pizza.Special.Description
         </div>
         <form class="dialog-body">
             <div>
                 <label>Size:</label>
-                <input type="range" min="@Pizza.MinimumSize" max="@Pizza.MaximumSize" step="1" @bind="Pizza.Size" />
-                <span class="size-label">@(Pizza.Size)" (£@(Pizza.GetFormattedTotalPrice()))</span>
+                <input type="range" min="@Pizza.MinimumSize" max="@Pizza.MaximumSize" step="1" @bind="Pizza.Size" @bind:event="oninput" />
+                <span class="size-label">
+                    @(Pizza.Size)" (£@(Pizza.GetFormattedTotalPrice()))
+                </span>
             </div>
         </form>
 
         <div class="dialog-buttons">
             <button class="btn btn-secondary mr-auto" @onclick="OnCancel">Cancel</button>
             <span class="mr-center">
-                Price:<span class="price">@(Pizza.GetFormattedTotalPrice())</span>
+                Price: <span class="price">@(Pizza.GetFormattedTotalPrice())</span>
             </span>
             <button class="btn btn-success ml-auto" @onclick="OnConfirm">Order ></button>
         </div>
@@ -25,7 +27,10 @@
 </div>
 
 @code {
-    [Parameter] public Pizza Pizza { get; set; }
-    [Parameter] public EventCallback OnCancel { get; set; }
-    [Parameter] public EventCallback OnConfirm { get; set; }
+    [Parameter, EditorRequired] 
+    public required Pizza Pizza { get; set; }
+    [Parameter, EditorRequired] 
+    public EventCallback OnCancel { get; set; }
+    [Parameter, EditorRequired] 
+    public EventCallback OnConfirm { get; set; }
 }
