@@ -3,11 +3,12 @@
     <p>
         <strong>
             @(pizza.Size)"
-            @pizza.Special.Name
+            @pizza.Special!.Name
             (£@pizza.GetFormattedTotalPrice())
         </strong>
     </p>
 }
+
 <p>
     <strong>
         Total price:
@@ -16,5 +17,6 @@
 </p>
 
 @code {
-    [Parameter] public Order Order { get; set; }
+    [Parameter, EditorRequired]
+    public required Order Order { get; set; }
 }
